+++ audfprint/__init__.py
@@ -1,0 +1,5 @@
+# coding=utf-8
+import logging
+
+LOG = logging.getLogger(__name__)
+LOG.addHandler(logging.NullHandler())
